slider images api add
