+++ src/context/Context.jsx
@@ -12,10 +12,9 @@ export function YosProvider({ children }) {
   const [universities, setUniversities] = useState([]);
   const [departments, setDepartments] = useState([]);
   const [card, setCard] = useState([]);
-  
+  const [images, setImages] = useState([]); // Yeni images state'i eklendi
 
   useEffect(() => {
-    
     axios
       .get(
         "https://tr-yös.com/api/v1/location/allcities.php?token=SX2qL5O3ivipPSMIWN8nXnaLWOiy4cEq7UdgZk448T5ZDpT1qbgMIrXVNquP1CWyNAH3JvoEVqnjiyg20a17549275a86d0e835660e56847e87a"
@@ -25,10 +24,9 @@ export function YosProvider({ children }) {
       })
       .catch((error) => {
         console.log(error);
-      }); 
+      });
 
-
-      axios
+    axios
       .get(
         "https://tr-yös.com/api/v1/education/alluniversities.php?token=SX2qL5O3ivipPSMIWN8nXnaLWOiy4cEq7UdgZk448T5ZDpT1qbgMIrXVNquP1CWyNAH3JvoEVqnjiyg20a17549275a86d0e835660e56847e87a"
       )
@@ -39,9 +37,9 @@ export function YosProvider({ children }) {
         console.log(error);
       });
 
-      axios
+    axios
       .get(
-          "https://tr-yös.com/api/v1/record/alldepartments.php?token=SX2qL5O3ivipPSMIWN8nXnaLWOiy4cEq7UdgZk448T5ZDpT1qbgMIrXVNquP1CWyNAH3JvoEVqnjiyg20a17549275a86d0e835660e56847e87a"
+        "https://tr-yös.com/api/v1/record/alldepartments.php?token=SX2qL5O3ivipPSMIWN8nXnaLWOiy4cEq7UdgZk448T5ZDpT1qbgMIrXVNquP1CWyNAH3JvoEVqnjiyg20a17549275a86d0e835660e56847e87a"
       )
       .then((response) => {
         setCard(response.data);
@@ -49,7 +47,6 @@ export function YosProvider({ children }) {
       .catch((error) => {
         console.log(error);
       });
- 
 
     axios
       .get(
@@ -61,6 +58,20 @@ export function YosProvider({ children }) {
       .catch((error) => {
         console.log(error);
       });
+
+    axios
+      .get("https://api.pexels.com/v1/search?query=university+education", {
+        headers: {
+          Authorization: "jSTPFVvp32B2JQtsEthjhm2ibITpE2OBuPjAoZaZiogsEsu66IzQodP0",
+        },
+      })
+      .then((response) => {
+        const imageUrls = response.data.photos.map((photo) => photo.src.large);
+        setImages(imageUrls);
+      })
+      .catch((error) => {
+        console.error(error);
+      });
   }, []);
 
   const contextValue = {
@@ -68,12 +79,8 @@ export function YosProvider({ children }) {
     universities,
     departments,
     card,
-   
+    images, // images state'i context değerlerine eklendi
   };
 
-  return (
-    <YosContext.Provider value={contextValue}>
-      {children}
-    </YosContext.Provider>
-  );
+  return <YosContext.Provider value={contextValue}>{children}</YosContext.Provider>;
 }
